auth_flask_app/requirements.txt
flask
authlib
python-dotenv